refactor: extract SplitViewLayout into subcomponents (#185)
Split 330-line SplitViewLayout into ResizablePanel, ResizableHandle,
ResizableContext, layout-persistence, and SplitViewLayout composition.
No behavior changes.
