+++ @fanslib/apps/web/src/components/SplitViewLayout/index.tsx
@@ -1,0 +1,1 @@
+export { SplitViewLayout } from "./SplitViewLayout";
