+++ @fanslib/apps/web/src/components/SplitViewLayout/layout-persistence.ts
@@ -1,0 +1,8 @@
+export const saveLayoutToLocalStorage = (id: string, layout: number[]) => {
+  localStorage.setItem(`resizable-layout-${id}`, JSON.stringify(layout));
+};
+
+export const getLayoutFromLocalStorage = (id: string): number[] | null => {
+  const layout = localStorage.getItem(`resizable-layout-${id}`);
+  return layout ? JSON.parse(layout) : null;
+};
